Toggle fullscreen directly in ShakaVideo
